% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.r
\name{fn_nnet}
\alias{fn_nnet}
\title{Simple package to build a Neural Network classifier for the Bank sales data.}
\usage{
fn_nnet()
}
\value{
Saved model file
}
\description{
Simple package to build a Neural Network classifier for the Bank sales data.
}
\examples{
fn_nnet()
}
