% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.r
\name{score}
\alias{score}
\title{Predict function to process the new data and return confusion matrix.}
\usage{
score(new_data)
}
\value{
confussion matrix table
}
\description{
Predict function to process the new data and return confusion matrix.
}
\examples{
score(new_data)
}
